changed result type to usize (because there would be integer overflow otherwise), solved day 7

+++ src/day_02/part1.rs
@@ -67,6 +67,6 @@ fn is_safe(report: &str, levels: &Vec<i32>) -> bool {
     true
 }
 
-pub fn part1(input: String) -> i32 {
-    count_safe_reports(input, is_safe)
+pub fn part1(input: String) -> usize {
+    count_safe_reports(input, is_safe) as usize
 }
