+++ src/day_16/part2.rs
@@ -1,3 +1,3 @@
-pub fn part2(input: String) -> i32 {
+pub fn part2(input: String) -> usize {
     0
 }
